requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
greenlet==3.3.2
groq==0.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.3
langchain-text-splitters==0.0.2
langchainhub==0.1.15
langsmith==0.1.17
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.7
packaging==23.2
propcache==0.4.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.3
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==8.5.0
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
